+++ test/integration/elements/test-element.ts
@@ -32,12 +32,20 @@ class TestElement extends Polymer.Element {
 
   @property({readOnly:true})
   readOnlyString: string;
+  
+  @property({computed:'computeString(reflectedString)'})
+  computedString: string;
 
+  @property({observer:'observeString'})
+  observedString: string;
+  
   // stand-in for set function dynamically created by Polymer on read only properties
   _setReadOnlyString: (value: string) => void;
 
   lastNumChange: number;
 
+  lastChange: string;
+  
   lastMultiChange: any[];
 
   @query('#num')
@@ -59,5 +67,13 @@ class TestElement extends Polymer.Element {
   @observe(['aNum', 'aString'])
   private _numStringChanged(newNum: number, newString: string) {
     this.lastMultiChange = [newNum, newString];
+  }
+  
+  private computeString(s:string) {
+      return "computed " + s;
+  }
+  
+  private observeString(s:string) {
+      this.lastChange = s;
   }
 }